add extra button for GPD Win5
L4 is mapped by default to LEFTCTRL + LEFTSHIFT + F14, but in WinControls, it only shows LEFTCTRL + LEFTSHIFT, so it can be inferred that F14 is included by default.
R4 is mapped by default to F3 + F15. Similarly, in WinControls, it only shows F3, so it can be inferred that F15 is included by default.